Introduce catch all mouse area
To be able to handle clicks outside the search box, we now have a
catch-all mouse area. This area does not accept any events but always
focuses the comment table. It also overrides all mouse cursors in the
main view.
The split view drag handle is the only other area where the user expects
a different mouse handle. Since listening to the hovered state change is
not officially supported, we introduce a workaround that emits a signal
whenever the mouse enters or leaves a split view drag handle.