Support embedding using hwnd and libmpv
Support in Qt is still preliminary. Hwnd embedding on Windows basically
is a massive hack. We stack three Windows on top of each other.
This is needed, because it is not possible to draw over the embedded
window if the elements to draw themselves are not in a separate window.
On Windows the architecture looks like this:

Main Window
  QtQuick Window
  Mpv Window