Use hWnd Embedding on Windows (#84)
Support in Qt is still preliminary. hWnd embedding on Windows basically
is a hack. We stack three Windows on top of each other. This is needed, 
because it is not possible to draw over the embedded window if the 
elements to draw themselves are not in a separate window.

On Windows the architecture looks like this:

Main Window
  Control Window
  Mpv Window

The user never interacts with the Main Window. The Control Window hosts
the entire user interface. It uses a transparent background. The actual
background color originates from the Main Window.

Currently, the window handle embedding is only done for Windows. On
Linux there's currently no window flag combination to get that stuff
working nicely. The video embedding itself works but the interaction 
with other windows on the desktop is broken (due to window stacking).