Apply settings changes only on confirmation (#104)
Previously, changes in appearance, backup, import, and export
settings dialogs were applied directly to the settings object.

Now, for appearance settings, we still apply changes directly
but reset them if the user does not click "Ok". For the other
three dialogs, changes are stored temporarily and applied only
when "Ok" is clicked.