feat: use uv to manage project (#120)
uv is new, shiny, and superfast. We can use it to manage virtual 
environments and dependencies. All in all, it's a better and widely
used tool to work with Python projects.

It also allows us to avoid using the internet when resolving latest and
pinned dependency versions. Finally, I used this PR to avoid installing
dev dependencies prior to building the Windows distributable.